requirements.txt
tensorflow-gpu==1.3.0
keras==2.0.5 
-e git+http://github.com/tensorflow/cleverhans.git#egg=cleverhans
